+++ src/WebLinter/Linters/CoffeeLinter.cs
@@ -5,7 +5,7 @@ namespace WebLinter
 {
     internal class CoffeeLinter : LinterBase
     {
-        private static Regex _rx = new Regex(@"^(?<file>[^,]+),(?<line>[0-9]+),(?<column>[0-9]+)?,.+,(?<message>[^,]+)$", RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Multiline);
+        private static Regex _rx = new Regex(@"^(?<file>.+),(?<line>[0-9]+),(?<column>[0-9]+)?,.+,(?<message>[^,\[]+)$", RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Multiline);
 
         public CoffeeLinter(ISettings settings, string workingDirectory) : base(settings, workingDirectory, _rx)
         {