+++ src/WebLinterVsix/Constants.cs
@@ -3,6 +3,5 @@
     class Constants
     {
         public const string VSIX_NAME = "Web Linter";
-        public const string PRODUCT_ID = "a9b5a26c-0774-412b-9b0e-8007614a9c73";
     }
 }